PF348 conserve le PRIS lors de l’invitation d’un opérateur

+++ app/models/projet.rb
@@ -183,13 +183,13 @@ class Projet < ActiveRecord::Base
     ProjetMailer.notification_invitation_intervenant(invitation).deliver_later!
     EvenementEnregistreurJob.perform_later(label: 'invitation_intervenant', projet: self, producteur: invitation)
 
-    if previous_operateur
+    if intervenant.operateur? && previous_operateur
       previous_invitation = invitations.where(intervenant: previous_operateur).first
       ProjetMailer.resiliation_operateur(previous_invitation).deliver_later!
       previous_invitation.destroy!
     end
 
-    if previous_pris
+    if intervenant.pris? && previous_pris
       previous_invitation = invitations.where(intervenant: previous_pris).first
       previous_invitation.destroy!
     end
